requirements.txt
djangorestframework==3.13.1
Django==4.1
flake8==5.0.4


